requirements.txt
langchain
langchain-community
langchain-gemini
faiss-cpu
sentence-transformers
pypdf
streamlit
langdetect
python-dotenv
